chore: idiomatic test helper exports
